Developer Build: Added some styling changes Added ability for sync'ing POEex rates Added setting for auto updating rates on login(default false) Disabled OneChoas recipe till there is a way of determining iLVL Fixed tabcontrol issue with league change
git-svn-id: http://procurement.googlecode.com/svn/trunk@43 f23b34ab-6a1b-74dd-ae4f-c8c00ee5c9be

+++ POEApi.Model/Currency.cs
@@ -3,7 +3,7 @@
     public class Currency : Item
     {
         public OrbType Type { get; private set; }
-        public float GCPValue { get; private set; }
+        public double GCPValue { get; private set; }
         public StackInfo StackInfo { get; private set; }
 
         public Currency(JSONProxy.Item item) : base(item)
