+++ POEApi.Model/CurrencyRatio.cs
@@ -3,10 +3,10 @@
     public class CurrencyRatio
     {
         public OrbType OrbType { get; set; }
-        public float OrbAmount { get; set; }
-        public float GCPAmount { get; set; }
+        public double OrbAmount { get; set; }
+        public double GCPAmount { get; set; }
 
-        public CurrencyRatio(OrbType orbType, int OrbAmount, int GCPAmount)
+        public CurrencyRatio(OrbType orbType, double OrbAmount, double GCPAmount)
         {
             this.OrbType = orbType;
             this.OrbAmount = OrbAmount;
